Improvements to the build system imported from other projects

+++ run.sh
@@ -3,16 +3,19 @@
 set -eu
 export LC_ALL=C
 
+DOCKER=$(command -v docker 2>/dev/null)
+
+IMAGE_REGISTRY=docker.io
 IMAGE_NAMESPACE=hectormolinero
 IMAGE_PROJECT=goaccess
 IMAGE_TAG=latest
-IMAGE_NAME=${IMAGE_NAMESPACE}/${IMAGE_PROJECT}:${IMAGE_TAG}
+IMAGE_NAME=${IMAGE_REGISTRY:?}/${IMAGE_NAMESPACE:?}/${IMAGE_PROJECT:?}:${IMAGE_TAG:?}
 
-imageExists() { [ -n "$(docker images -q "$1")" ]; }
+imageExists() { [ -n "$("${DOCKER:?}" images -q "${1:?}")" ]; }
 
-if ! imageExists "${IMAGE_NAME}"; then
-	>&2 printf -- '%s\n' "\"${IMAGE_NAME}\" image doesn't exist!"
+if ! imageExists "${IMAGE_NAME:?}" && ! imageExists "${IMAGE_NAME#docker.io/}"; then
+	>&2 printf -- '%s\n' "\"${IMAGE_NAME:?}\" image doesn't exist!"
 	exit 1
 fi
 
-exec docker run -it --rm "${IMAGE_NAME}" "$@"
+exec docker run -it --rm "${IMAGE_NAME:?}" "$@"
